knowledge-base/requirements.txt
fastapi
uvicorn
python-multipart
lancedb
pypdf
python-docx
python-dotenv
pydantic
langchain
langchain-community
cohere
tantivy
pytest
httpx
pytest-asyncio
langchain-cohere
langchain-experimental
langchain_text_splitters
